libinterp/math: fix export_int/import_int for 64-bit Limbo int
math->export_int and math->import_int silently mangled every array of
more than one element. Round-tripping [1,2,3,4] gave [1,2,0,0], and
export produced 00000001 00000000 00000002 00000000.

Limbo's int is WORD, and this fork typedefs WORD as intptr - 64 bits
here, with the original "typedef int WORD; /* 32 bits */" left
commented out in include/interp.h. Both functions still cast the array
data to C int*, so they walked it in 4-byte steps: element 1 came back
as the top half of element 0, which for small positive values is zero.
It has presumably been broken since the 64-bit port; nothing in the
tree used these on a multi-element array until now.

Fixed by reading and writing through WORD*. The external form is
unchanged and still 32-bit big-endian - only the in-memory element type
was wrong - so any data already written by the old code on a 32-bit
build still reads correctly.

Found while giving gpu(3) a binary wire format, where a two-element
header came out with its second field zeroed. Verified directly:
export_int([1,2,3,4]) now yields 00000001 00000002 00000003 00000004
and import_int round-trips it. keyuptest, nowmtest, line3test,
temple/numericstest and a pde(2) solve all still pass.

+++ libinterp/math.c
@@ -36,20 +36,27 @@ Math_import_int(void *fp)
 	int i, n;
 	unsigned int u;
 	unsigned char *bp;
-	int *x;
+	WORD *x;
 
 	f = fp;
 	n = f->x->len;
 	if(f->b->len!=4*n)
 		error(exMathia);
 	bp = (unsigned char *)(f->b->data);
-	x = (int*)(f->x->data);
+	/*
+	 * Limbo's int is WORD, which is intptr and so 64 bits here; this
+	 * read as C int and silently mangled every array of more than one
+	 * element (element 1 came back as the top half of element 0).  The
+	 * external form stays 32-bit - only the in-memory element type
+	 * changed with the 64-bit port.
+	 */
+	x = (WORD*)(f->x->data);
 	for(i=0; i<n; i++){
 		u = *bp++;
 		u = (u<<8) | *bp++;
 		u = (u<<8) | *bp++;
 		u = (u<<8) | *bp++;
-		x[i] = u;
+		x[i] = (int)u;	/* sign-extend the 32-bit external value */
 	}
 }
 
@@ -114,16 +121,17 @@ Math_export_int(void *fp)
 	int i, n;
 	unsigned int u;
 	unsigned char *bp;
-	int *x;
+	WORD *x;
 
 	f = fp;
 	n = f->x->len;
 	if(f->b->len!=4*n)
 		error(exMathia);
 	bp = (unsigned char *)(f->b->data);
-	x = (int*)(f->x->data);
+	/* See Math_import_int: Limbo int is WORD (64-bit here), not C int. */
+	x = (WORD*)(f->x->data);
 	for(i=0; i<n; i++){
-		u = x[i];
+		u = (unsigned int)x[i];
 		*bp++ = u>>24;
 		*bp++ = u>>16;
 		*bp++ = u>>8;
